# docs/swagger/swagger.yaml
basePath: /
definitions:
  domain.ApiKey:
    properties:
      created_at:
        type: string
      id:
        type: string
      key:
        description: 'Transparently: the actual secret'
        type: string
      last_used:
        type: string
      name:
        type: string
      user_id:
        type: string
    type: object
  domain.DashboardStats:
    properties:
      cpu_history:
        items:
          $ref: '#/definitions/domain.MetricPoint'
        type: array
      memory_history:
        items:
          $ref: '#/definitions/domain.MetricPoint'
        type: array
      recent_events:
        items:
          $ref: '#/definitions/domain.Event'
        type: array
      summary:
        $ref: '#/definitions/domain.ResourceSummary'
    type: object
  domain.Event:
    properties:
      action:
        description: e.g. INSTANCE_LAUNCH
        type: string
      created_at:
        type: string
      id:
        type: string
      metadata:
        description: JSON details
        items:
          type: integer
        type: array
      resource_id:
        description: e.g. UUID of instance
        type: string
      resource_type:
        description: e.g. INSTANCE, VPC
        type: string
    type: object
...
  domain.MetricPoint:
    properties:
      label:
        type: string
      timestamp:
        type: string
      value:
        type: number
    type: object
...
  domain.ResourceSummary:
    properties:
      attached_volumes:
        type: integer
      running_instances:
        type: integer
      stopped_instances:
        type: integer
      total_instances:
        type: integer
      total_storage_mb:
        type: integer
      total_volumes:
        type: integer
      total_vpcs:
        type: integer
    type: object
  domain.ScalingGroup:
    properties:
      created_at:
        type: string
      current_count:
        type: integer
      desired_count:
        type: integer
      failure_count:
        type: integer
      id:
        type: string
      idempotency_key:
        type: string
      image:
        type: string
      last_failure_at:
        type: string
      load_balancer_id:
        type: string
      max_instances:
        type: integer
      min_instances:
        type: integer
      name:
        type: string
      ports:
        type: string
      status:
        $ref: '#/definitions/domain.ScalingGroupStatus'
      updated_at:
        type: string
      version:
        type: integer
      vpc_id:
        type: string
    type: object
  domain.ScalingGroupStatus:
    enum:
    - ACTIVE
    - UPDATING
    - DELETING
    - DELETED
    type: string
    x-enum-varnames:
    - ScalingGroupStatusActive
    - ScalingGroupStatusUpdating
    - ScalingGroupStatusDeleting
    - ScalingGroupStatusDeleted
  domain.ScalingPolicy:
    properties:
      cooldown_sec:
        type: integer
      id:
        type: string
      last_scaled_at:
        type: string
      metric_type:
        description: '"cpu" | "memory"'
        type: string
      name:
        type: string
      scale_in_step:
        type: integer
      scale_out_step:
        type: integer
      scaling_group_id:
        type: string
      target_value:
        type: number
    type: object
...
  httphandlers.CreateASPolicyRequest:
    properties:
      cooldown_sec:
        type: integer
      metric_type:
        type: string
      name:
        type: string
      scale_in_step:
        type: integer
      scale_out_step:
        type: integer
      target_value:
        type: number
    required:
    - cooldown_sec
    - metric_type
    - name
    - scale_in_step
    - scale_out_step
    - target_value
    type: object
  httphandlers.CreateGroupRequest:
    properties:
      desired_count:
        type: integer
      image:
        type: string
      load_balancer_id:
        type: string
      max_instances:
        type: integer
      min_instances:
        description: 0 is valid
        type: integer
      name:
        type: string
      ports:
        type: string
      vpc_id:
        type: string
    required:
    - desired_count
    - image
    - max_instances
    - name
    - vpc_id
    type: object
  httphandlers.CreateKeyRequest:
    properties:
      name:
        type: string
    required:
    - name
    type: object
...
  httputil.Meta:
    properties:
      request_id:
        type: string
      timestamp:
        type: string
    type: object
  httputil.Response:
    properties:
      data: {}
      error: {}
      meta:
        $ref: '#/definitions/httputil.Meta'
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is the Mini AWS Compute API server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Mini AWS API
  version: "1.0"
paths:
  /api/dashboard/events:
    get:
      description: Gets a list of the latest audit/status events
      parameters:
      - default: 10
        description: Number of events to return
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Event'
            type: array
      security:
      - ApiKeyAuth: []
      summary: Get recent events
      tags:
      - dashboard
  /api/dashboard/stats:
    get:
      description: Gets comprehensive counts and aggregated resource information
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.DashboardStats'
      security:
      - ApiKeyAuth: []
      summary: Get detailed dashboard stats
      tags:
      - dashboard
  /api/dashboard/stream:
    get:
      description: Real-time stream of dashboard summary updates via Server-Sent Events
      produces:
      - text/event-stream
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: Stream dashboard updates (SSE)
      tags:
      - dashboard
  /api/dashboard/summary:
    get:
      description: Gets counts for instances, volumes, vpcs, and events
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.ResourceSummary'
      security:
      - ApiKeyAuth: []
      summary: Get dashboard summary
      tags:
      - dashboard
  /auth/keys:
    post:
      consumes:
      - application/json
      description: Bootstraps access by generating an API key for a given name
      parameters:
      - description: Key creation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httphandlers.CreateKeyRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.ApiKey'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.Response'
      summary: Create a new API key
      tags:
      - identity
  /autoscaling/groups:
    get:
      description: Gets a list of all auto-scaling groups
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.ScalingGroup'
            type: array
      security:
      - ApiKeyAuth: []
      summary: List all scaling groups
      tags:
      - autoscaling
    post:
      consumes:
      - application/json
      description: Creates an auto-scaling group for instances
      parameters:
      - description: ASG creation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httphandlers.CreateGroupRequest'
      - description: Idempotency key
        in: header
        name: Idempotency-Key
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.ScalingGroup'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.Response'
      security:
      - ApiKeyAuth: []
      summary: Create a new scaling group
      tags:
      - autoscaling
  /autoscaling/groups/{id}:
    delete:
      description: Removes an auto-scaling group
      parameters:
      - description: ASG ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httputil.Response'
      security:
      - ApiKeyAuth: []
      summary: Delete a scaling group
      tags:
      - autoscaling
    get:
      description: Gets detailed information about a specific auto-scaling group
      parameters:
      - description: ASG ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.ScalingGroup'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httputil.Response'
      security:
      - ApiKeyAuth: []
      summary: Get scaling group details
      tags:
      - autoscaling
  /autoscaling/groups/{id}/policies:
    post:
      consumes:
      - application/json
      description: Adds a scaling policy to an auto-scaling group
      parameters:
      - description: ASG ID
        in: path
        name: id
        required: true
        type: string
      - description: Policy creation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httphandlers.CreateASPolicyRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.ScalingPolicy'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.Response'
      security:
      - ApiKeyAuth: []
      summary: Create a new scaling policy
      tags:
      - autoscaling
  /autoscaling/policies/{id}:
    delete:
      description: Removes a scaling policy
      parameters:
      - description: Policy ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httputil.Response'
      security:
      - ApiKeyAuth: []
      summary: Delete a scaling policy
      tags:
      - autoscaling
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: X-API-Key
    type: apiKey
swagger: "2.0"